DEcompare

Function that takes a list of 2 dataframes with expression data and compares the log2FoldChange values of the two data sets.

DEcompare(
  deg_list,
  threshold = 1.5,
  genes = NULL,
  xlim = c(-10, 10),
  ylim = c(-10, 10),
  xlab = "Contrast1",
  ylab = "Contrast2",
  main = "Comparison of Log2FC",
  subtitle = paste(xlab, "vs", ylab),
  color_corners = c("pink", "lightgreen", "cornflowerblue", "yellow"),
  alpha_corners = c(0.7)
)

Arguments

deg_list
List of lenght 2. Must contain data frames with the columns Geneid, log2FoldChange, padj and DEG. Values in 'Geneid' must be the same in both data frames.+

threshold
Numeric of lenght 1. Threshold for the log2FoldChange. It is used to draw horizontal and vertical lines. Default: 1.5.

genes
Charachter of indefined lenght or NULL. Genes to be written in the plot. Default: NULL.

xlim
Numeric of length 2. The limits of the Y axis. Default: c(-10, 10)

xlab
Character of lenght 1. Name of the X axis. Default: "Contrast1"

ylab
Character of lenght 1. Name of the Y axis. Default: "Contrast2"

subtitle
Character of lenght 1. Subtitle of the plot. Default: paste(xlab, "vs", ylab)

alpha_corners
Numeric of length 1 to 4. If too short, it will be recycled. Alpha (transparency) value for the vars. Default: 0.7

title
Character of lenght 1. Title of the plot. Default: "Comparison of Log2FC"

colors
Character of length 1 to 4. If too short, it will be recycled. Colors for each of the corners: top-left, bottom-left, top-right, bottom-right. Default: c("pink", "lightgreen", "cornflowerblue", "yellow")

Author

amitjavilaventura