ggVennBed

Does the intersection of two BED files (A and B) and draws the intersection with the unique peaks of A and B and the intersections of A with B and B with A.

ggVennBed(
  a,
  b,
  setnames = c("A", "B"),
  stranded = T,
  color = c("blue", "gold3"),
  namesize = 7,
  labsize = 5
)

Arguments

a

Data frame or character of length 1. Path to a BED file or data frame with BED format. If data frame, it will be written to a temporary file.

b

Data frame or character of length 1. Path to a BED file or data frame with BED format. If data frame, it will be written to a temporary file.

setnames

Character of length 2. Names of the data sets of 'a' and 'b'. Default: c("A", "B").

stranded

Logical of length 1. If TRUE, the intersections are done regarding the strand. Default = TRUE.

color

Character of length 2. Color of the Venn circles and the labels indicating the number of regions. Default: c("blue", "gold3").

namesize

Numeric of length 1. Size of the names of each set. Default: 7.

labsize

Numeric of length 1. Size of the labels indicating the number of unique and overlapping regions.

See also

Author

amitjavilaventura