plotDendogram

Helper function that takes clustering data (hclust()) as input and plots a dendogram using ggplot and ggdendro

plotDendogram(hclust_data, axis = "rows")

Arguments

hclust_data
Output from hclust() function.

axis
Characteer of length 1. Either "rows" or "cols". If rows, the dendogram is made to be placed in the left of the Y axis; if cols, the dendogram is made to be placed in the top of the X axis.

See also

Author

amitjavilaventura